Feature key code mapper and modifier key enum
- Add `KeyCodeMapper`
- Add `ModifierKey`
- Add `InputerController` & `InputSource`

The `KeyCodeMapper` is used to map key code back to a string.
This can be used when trying to translate keyboard invocation into
`Combination`, which is a string-based key identifier.

`ModifierKey` supplements the `KeyCodeMapper` by both being able to
pretty print values and translate the different modifier types.

When input comes in, we can filter the combination string by first
determining the modifiers and finally figuring out which key code
it translates into.

`InputSource` is used to correctly translate key codes into the current
selected input source.

`InputSourceController` is an abstraction in top of the Carbon-based
framework that is used inside `KeyCodeMapper`. The key code translation
needs to respect the current input source in order to produce correct
output, otherwise key bindings won't end up correct depending on the
users current input source.

`InputSourceController` is also used inside the `KeyCodeMapper` tests
to switch between different sources. Without this, then the tests would
fail on CI and other developer machines if the select input didn't match
the input of the tests results.