Feature basic drop application support
Add basic drag and drop applications into the list of groups.
When dragging an application into the group list, it will create a new group with the same name
as the application (using the bundleName) and it adds a basic rule that will scope all workflows
inside of the that group to the dropped application.

- Add `installedApplications` to `GroupFeatureController`
- Add `DispatchQueue.main` to `GroupFeatureController.reload`
- Expose installed applications on `CoreControlling`
- Add new action on `GroupList` for dropped applications
- Improve `GroupList` by extracting the list into a variable