Send ids to the re-render method, this fixes an issue where if switch groups fast, it could get stuck on the previous selection. There is still more work to be done here because not all the issues are fixed. Added a note about the `guard` that shouldn't be there.
We should also consider removing the subscriptions and instead have a more
explicit workflow by listening directly to invocations from the actions
coming from the views.

Perhaps not all subscription should go away, but we should aim to have a
system that doesn't sole rely on the side-effects of Combine subscriptions
as that make the flow a lot harder to understand.