Refactor Command data structure
Move common variables to `Command.MetaData`.
These values include: `delay`, `id`, `name`, `isEnabled` and `notification`.
By unifying the data structure across all the different commands, it will
make it easier to work with and some parts of the code can becomes totally
agnostic to the actual command type and just deal with the meta data.

For example, we won't have to unwrap the command to figure out any of the
underlying variable, we can simply access the meta data directly.
This makes code a lot more cleaner and we get rid of unwrapping and casting
which can lead to some performance gains.