Improve the process of giving the application permissions
On the first launch of Keyboard Cowboy, it will enter a mode that checks for permissions every time the frontmost application changes.
This process will continue until the app has obtained the necessary permissions to function correctly.
Once this occurs, the subscription will be invalidated, and a working mach port will be set up.
This eliminates the need to restart the application after obtaining permissions, resulting in a significant improvement to the user experience.