Refactor core setup of the application
Add new `Core` class that holds all the dependecies of the application.
It cleans up the `KeyboardCowboy` application struct and makes it easier
to ensure that we don't create more class instances than we need.

With a bit more work, it should be easier to test the application.

This change also cleans up some SwiftUI related components which
should bring some slight performance gains because the views no
longer will respond to "bogus" updates.