Refactor Finder and Shortcut classes to generalize naming
We have updated the code to generalize the naming convention of certain classes in the scripting bridge, replacing the specific 'Finder' prefix with a more general 'SB' prefix to improve the reusability of the code. In addition, we've implemented an async version of the shortcut indexing method and added subscription logic to the ShortcutStore to react to changes in the front-most application, specifically for the "com.apple.shortcuts" bundle identifier. This allows the ShortcutStore to dynamically update the list of shortcuts when the shortcuts application becomes active, providing a more responsive user experience.