Improved resolving frame in accessibility element resolvers to return CGRect directly
In our accessibility element resolution strategy, we have refactored multiple resolvers such as AXCollectionResolver, AXGroupResolver, AXListResolver, AXOutlineResolver, AXScrollAreaResolver, and introduced AXTextEntryAreaResolver to directly return the frame (CGRect) of the focused element instead of returning an AnyAccessibilityElement. Previously, we were obtaining an AnyAccessibilityElement and then extracting its frame which was an unnecessary intermediate step. This change simplifies the API and directly provides the essential data needed for further processing. Additionally, the MouseCommandRunner has been updated to handle the new return type and to consider the newly added case .textEntryArea while resolving the frame. The added error case in MouseCommandRunnerError ensures that an appropriate error is thrown when a frame cannot be resolved, enhancing error handling.