Refactor click location handling in MouseCommand model and XcodeMouseRoutine
To improve the maintainability and extendibility of the MouseCommand model, we've refactored the way click locations are accessed and modified. The MouseCommand.UIElement enum now includes a computed property for directly getting and setting the ClickLocation without duplicating switch-case logic. Additionally, the XcodeMouseRoutine has been updated to utilize this new property, streamlining the logic for frame calculations and making it easier to add new click location behaviors in the future. This change promotes cleaner code practices and prepares for potential feature expansion.